+++ pydidas/gui/frames/workflow_test_frame.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as published by
+# the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -272,7 +274,7 @@ class WorkflowTestFrame(WorkflowTestFrameBuilder):
         Reload the local WorkflowTree from the global one, e.g. to propagate changes
         to global settings.
         """
-        self._tree = TREE.get_copy()
+        self._tree = TREE.deepcopy()
         self.param_widgets["selected_results"].setCurrentIndex(0)
         self._config["plugin_res_titles"] = {}
         self.__update_selection_choices()