+++ pydidas/plugins/base_input_plugin.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as published by
+# the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -48,7 +50,7 @@ class InputPlugin(BasePlugin):
     output_data_label = "Image intensity"
     output_data_unit = "counts"
     input_data_dim = 2
-    generic_params = BasePlugin.generic_params.get_copy()
+    generic_params = BasePlugin.generic_params.copy()
     generic_params.add_params(
         get_generic_parameter("use_roi"),
         get_generic_parameter("roi_xlow"),
@@ -58,7 +60,7 @@ class InputPlugin(BasePlugin):
         get_generic_parameter("binning"),
         get_generic_parameter("live_processing"),
     )
-    default_params = BasePlugin.default_params.get_copy()
+    default_params = BasePlugin.default_params.copy()
 
     def __init__(self, *args, **kwargs):
         """