+++ pydidas/plugins/base_input_plugin_1d.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as published by
+# the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -44,14 +46,14 @@ class InputPlugin1d(InputPlugin):
     plugin_type = INPUT_PLUGIN
     plugin_name = "Base input plugin 1d"
     input_data_dim = 1
-    generic_params = BasePlugin.generic_params.get_copy()
+    generic_params = BasePlugin.generic_params.copy()
     generic_params.add_params(
         get_generic_parameter("use_roi"),
         get_generic_parameter("roi_xlow"),
         get_generic_parameter("roi_xhigh"),
         get_generic_parameter("binning"),
     )
-    default_params = InputPlugin.default_params.get_copy()
+    default_params = InputPlugin.default_params.copy()
 
     def __init__(self, *args, **kwargs):
         """