+++ pydidas/plugins/base_plugin.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as published by
+# the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -203,7 +205,7 @@ class BasePlugin(ObjectWithParameterCollection):
             OUTPUT_PLUGIN,
         ]:
             raise ValueError("Unknown value for the plugin type")
-        self.add_params(self.generic_params.get_copy(), *args)
+        self.add_params(self.generic_params.copy(), *args)
         self.set_default_params()
         self.update_param_values_from_kwargs(**kwargs)
         for _kw, _item in kwargs.items():