+++ pydidas/plugins/base_proc_plugin.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as published by
+# the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -36,5 +38,5 @@ class ProcPlugin(BasePlugin):
     plugin_type = PROC_PLUGIN
     plugin_subtype = PROC_PLUGIN_GENERIC
     plugin_name = "Base processing plugin"
-    generic_params = BasePlugin.generic_params.get_copy()
-    default_params = BasePlugin.default_params.get_copy()
+    generic_params = BasePlugin.generic_params.copy()
+    default_params = BasePlugin.default_params.copy()