+++ pydidas/widgets/base_frame_with_app.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as published by
+# the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -74,7 +76,7 @@ class BaseFrameWithApp(BaseFrame):
         if not isinstance(app, BaseApp):
             raise TypeError("The passed object must be a BaseApp instance.")
         if not isinstance(self._app, BaseApp):
-            self._app = app.get_copy()
+            self._app = app.copy()
             self._app.slave_mode = False
             return
         for param_key in app.params: