+++ pydidas_plugins/input_plugins/fio_mca_line_scan_series_loader.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as published by
+# the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -132,7 +134,7 @@ class FioMcaLineScanSeriesLoader(InputPlugin1d):
     plugin_name = "Fio MCA line scan series loader"
     basic_plugin = False
     plugin_type = INPUT_PLUGIN
-    default_params = FIO_MCA_READER_DEFAULT_PARAMS.get_copy()
+    default_params = FIO_MCA_READER_DEFAULT_PARAMS.copy()
     input_data_dim = None
     output_data_dim = 1