+++ pydidas_plugins/proc_plugins/subtract_1d_bg_profile.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as published by
+# the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -72,7 +74,7 @@ class Subtract1dBackgroundProfile(ProcPlugin):
     plugin_type = PROC_PLUGIN
     plugin_subtype = PROC_PLUGIN_INTEGRATED
     default_params = get_generic_param_collection("process_data_dim")
-    default_params.add_params(_PARAM_THRESH.get_copy(), _PARAM_PROFILE_FILE.get_copy())
+    default_params.add_params(_PARAM_THRESH.copy(), _PARAM_PROFILE_FILE.copy())
     input_data_dim = 1
     output_data_dim = 1
     output_data_label = "Background-corrected data"