+++ tests/plugins/test_base_plugin.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as published by
+# the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -404,7 +406,7 @@ class TestBasePlugin(unittest.TestCase):
         }
         for _key, _val in _new_params.items():
             plugin.add_param(Parameter(_key, str, ""))
-        _state = {"params": plugin.params.get_copy()}
+        _state = {"params": plugin.params.copy()}
         for _key, _param in _new_params.items():
             _state["params"][_key].value = _new_params[_key]
         plugin.__setstate__(_state)