+++ tests/workflow/test_generic_node.py
@@ -1,9 +1,11 @@
 # This file is part of pydidas.
 #
+# Copyright 2021-, Helmholtz-Zentrum Hereon
+# SPDX-License-Identifier: GPL-3.0-only
+#
 # pydidas is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
+# it under the terms of the GNU General Public License version 3 as published by
+# the Free Software Foundation.
 #
 # Pydidas is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -324,11 +326,11 @@ class TestGenericNode(unittest.TestCase):
         self.assertNotEqual(root, root_copy)
         self.assertNotEqual(root._children, root_copy._children)
 
-    def test_get_copy(self):
+    def test_copy(self):
         root = GenericNode(node_id=0)
         GenericNode(node_id=1, parent=root)
         GenericNode(node_id=2, parent=root)
-        root_copy = root.get_copy()
+        root_copy = root.copy()
         self.assertNotEqual(root, root_copy)
         self.assertNotEqual(root._children, root_copy._children)